ToolCallSuccess tests: fold short-circuit unit tests into behavior file
Per PR review: asset.yaml's pytest.tests_dir points at a single file (test_tool_call_success_evaluator_behavior.py), so the standalone test_tool_call_success_status_shortcircuit.py was not being picked up by the asset's CI run. Move both unit-test classes (TestCollectFailedToolCalls, TestGetToolCallsResultsNoStatusForward) plus their helpers into the behavior file alongside the existing TestToolCallSuccessEvaluatorBehavior class, and delete the standalone file. 19 unit tests collected; pytest collects 88 total tests for tool_call_success.
